Widen storage layer to support log segments larger than 2GB
Phase 2 of the log.segment.bytes INT-to-LONG migration: widen the
storage layer internals so that physical file positions and segment
sizes can exceed Integer.MAX_VALUE.

FileRecords:
- Internal size tracking: AtomicInteger -> AtomicLong
- Added sizeInBytesLong(), sliceLong(), truncateToLong() for long-aware
  callers while keeping int-based interface methods for BaseRecords
  compatibility
- Widened start/end fields, open() initFileSize, position params to long
- Removed the Integer.MAX_VALUE constructor check
- LogOffsetPosition.position: int -> long

OffsetIndex (.index format):
- ENTRY_SIZE: 8 -> 12 bytes (4-byte relative offset + 8-byte position)
- append(): putInt(position) -> putLong(position)
- physical(): getInt() -> getLong()

OffsetPosition: int position -> long position

LogSegment:
- Added sizeInBytesLong() using FileRecords.sizeInBytesLong()
- shouldRoll() uses long for size comparison
- append() physicalPosition: int -> long
- recover() validBytes: int -> long
- open() initFileSize: int -> long

Cascading changes:
- LogOffsetMetadata.relativePositionInSegment: int -> long
- LogConfig.initFileSize(): int -> long
- SegmentPosition.relativePosition: int -> long
- Cleaner, LocalLog, LogLoader, UnifiedLog, RemoteIndexCache,
  RemoteLogManager: updated to use long positions

+++ clients/src/main/java/org/apache/kafka/common/record/internal/AbstractLegacyRecordBatch.java
@@ -546,7 +546,7 @@ public abstract class AbstractLegacyRecordBatch extends AbstractRecordBatch impl
         LegacyFileChannelRecordBatch(long offset,
                                      byte magic,
                                      FileRecords fileRecords,
-                                     int position,
+                                     long position,
                                      int batchSize) {
             super(offset, magic, fileRecords, position, batchSize);
         }
